Extracted class ProjectUpdater from UpdaterServiceImpl
